backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.0.3; python_version >= '3.10'
django-autoslug==1.9.9
django-cors-headers==4.3.1; python_version >= '3.8'
django-debug-toolbar==4.3.0; python_version >= '3.8'
django-filter==24.1; python_version >= '3.8'
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
drf-nested-routers==0.93.5; python_version >= '3.8'
gunicorn==21.2.0; python_version >= '3.5'
mysql-connector-python==8.3.0; python_version >= '3.8'
packaging==24.0; python_version >= '3.7'
pillow==10.2.0; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.10.0; python_version < '3.11'
whitenoise==6.6.0; python_version >= '3.8'
